Change habit summaries to render and send images
This commit squashes the below 8 changes into one:

1. Change image rendering library to imgkit

Besides adding the package itself in the package file, we also need to
install wkhtmltopdf in the container (imgkit is essentially a wrapper
around this from what I've heard).

2. Add dependency for imgkit in Dockerfile

3. Render summary tables using images

This commit changes tabulate to render an HTML table instead of ASCII,
and will use imgkit to render that into an image file. As a result, the
sent message had to be modified to send the image instead.

For cleanup purposes, images are deleted once sent.

4. Add style sheet for habit tracker summary tables

5. Include assets directory in Docker image

This change is required due to imgkit not taking the stylesheet as a
URL. The alternative to this would be to make a request to the github
url to get the static file from there instead, but for now I avoided
such a network call just for a CSS file that will change infrequently.

6. Update function declarations and docs for habit summary generation

7. Add missing os import

8. Format habit_tracker_cog.py with black

Improve habit tracker summary messages

This commit adds the header text from the previous messages back, so
that we know what habit is the send message referring to, and which
table is the weekly vs the total summary.

It also adds the accidentally removed function call that actually
generates the images used.

Improve habit tracker summary messages

This commit adds the header text from the previous messages back, so
that we know what habit is the send message referring to, and which
table is the weekly vs the total summary.

It also adds the accidentally removed function call that actually
generates the images used.